Move PendingMove declaration to tsanOopMapTable.hpp.
